requirements.txt
asgiref==3.3.4
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.4
django-environ==0.7.0
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
djoser==2.1.0
environ==1.0
gunicorn==20.0.4
drf-yasg==1.20.0
idna==2.10
importlib-metadata==1.7.0
inflection==0.5.1
iniconfig==1.1.1
isort==5.9.0
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
oauthlib==3.1.1
packaging==20.9
pluggy==0.13.1
psycopg2-binary==2.9.1
py==1.10.0
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
python-dotenv==0.19.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.5
zipp==3.4.1
